renamed project to portable
