+++ .gitignore
@@ -1,9 +1,21 @@
-**/obj/*
-**/Bin/*
-**/node_modules/*
+*.nupkg
 **/packages/*
-**/Components/*
-.nuget/*
-Components/*
-**/**/Bin/*
-**/**/obj/*
+!**/packages/build/
+
+*.suo
+*.user
+*.userosscache
+*.sln.docstates
+
+[Dd]ebug/
+[Dd]ebugPublic/
+[Rr]elease/
+[Rr]eleases/
+x64/
+x86/
+build/
+bld/
+[Bb]in/
+[Oo]bj/
+
+.vs/
