+++ FHSDKPortable/Properties/AssemblyInfo.cs
