+++ FHSDKPortable/Services/Device/DeviceService.cs
