+++ FHSDKPortable/Services/Hash/HashService.cs
