+++ FHSDKPortable/Services/Log/LogService.cs
