+++ FHSDKPortable/Services/Monitor/MonitorService.cs
