+++ FHSDKPortable/packages.config
